refactor: move search bar to components from routes
